@@ -1,20 +1,6 @@
-'use strict';
-
 import {
-  AppRegistry,
-  BackAndroid,
+  AppRegistry
 } from 'react-native';
-import Finance from './Finance';
-
-import { Actions } from 'react-native-router-flux';
-
-BackAndroid.addEventListener('hardwareBackPress', () => {
-  try {
-    Actions.pop();
-    return true;
-  } catch (err) {
-    return false;
-  }
-});
+import Finance from './finance';
 
 AppRegistry.registerComponent('Finance', () => Finance);
